refactor/restructure
all packages no encapsulated within rpm package; added more structure to
inner packages (security split into config and util), split models into
entities and dtos, put SwaggerConfig and AuthAsepct into rpm.util
package. more changes/file moving as well, less important
