+++ src/main/java/com/revature/rpm/dtos/UserErrorResponse.java
@@ -1,4 +1,4 @@
-package com.revature.models;
+package com.revature.rpm.dtos;
 
 import java.io.Serializable;
 import java.util.Objects;
