+++ src/main/java/com/revature/rpm/dtos/UserPrincipal.java
@@ -1,9 +1,11 @@
-package com.revature.models;
+package com.revature.rpm.dtos;
 
 import java.util.Collection;
 
 import org.springframework.security.core.GrantedAuthority;
 import org.springframework.security.core.userdetails.User;
+
+import com.revature.rpm.entities.AppUser;
 
 
 /**
