+++ src/main/resources/bootstrap.yml
@@ -22,4 +22,4 @@ spring:
   cloud:
     config:
       uri:
-      - http://localhost:8888
+      - http://localhost:8889
